trbajo

+++ Producto/Api-Horarios/Lambda_EliminarHorario.js
@@ -9,14 +9,13 @@ const FUNCION_VALIDAR = process.env.FUNCION_VALIDAR;
 exports.handler = async (event) => {
   try {
     const token = event.headers?.Authorization;
-    if (!token) return { statusCode: 403, body: JSON.stringify({ error: 'Token no proporcionado' }) };
+    const { tenant_id,curso_id, horario_id } = JSON.parse(event.body);
+    if (!token || !tenant_id) return { statusCode: 403, body: JSON.stringify({ error: 'Token o tenant_id no proporcionado' }) };
 
-    const validar = await lambda.invoke({ FunctionName: FUNCION_VALIDAR, InvocationType: 'RequestResponse', Payload: JSON.stringify({ token }) }).promise();
+    const validar = await lambda.invoke({ FunctionName: FUNCION_VALIDAR, InvocationType: 'RequestResponse', Payload: JSON.stringify({ token,tenant_id }) }).promise();
     const validarPayload = JSON.parse(validar.Payload);
     if (validarPayload.statusCode === 403) return { statusCode: 403, body: JSON.stringify({ error: 'Token inválido' }) };
 
-    const { tenant_id } = validarPayload.body;
-    const { curso_id, horario_id } = JSON.parse(event.body);
     const tenant_id$curso_id = `${tenant_id}#${curso_id}`;
 
     await dynamodb.delete({ TableName: TABLE_HORARIO, Key: { tenant_id$curso_id, horario_id } }).promise();
